device: build `preloader_device` package
Even though we are non-A/B, create_pl_dev is still needed on MTK because the preloader isnt a normal partition but hidden in the boot regions /dev/block/sda(b) or at mmcblk0boot0, create_pl_dev  makes device mapper devices (preloader_raw_a/b) that skip the MTK header

So: Just because you don't see it in the firmware dump, doesn't mean its not there, or isnt needed. Non-A/B does not matter the preloader is still there. Without it, OTA or fastbootd cant  find the  preloader.

Here you incorrectly dropped the mali suffix. 
hardware/mediatek/aidl/memtrack/Android.bp
cc_binary {
    name: "android.hardware.memtrack-service.mediatek-mali",
    relative_install_path: "hw",

poweroffalarm: maybe, but one, nobody actually ships that framework in a samsung device rom. And for second, we are looking at this device more so as a mediatek device than samsung, in terms of that is the hardware. And always build components from source, when you can.